+++ bin/fuel-core/src/cli/run.rs
@@ -500,6 +500,7 @@ impl Command {
             graphql_config: GraphQLConfig {
                 addr,
                 number_of_threads: graphql.graphql_number_of_threads,
+                database_batch_size: graphql.database_batch_size,
                 max_queries_depth: graphql.graphql_max_depth,
                 max_queries_complexity: graphql.graphql_max_complexity,
                 max_queries_recursive_depth: graphql.graphql_max_recursive_depth,