+++ bin/fuel-core/src/cli/run/graphql.rs
@@ -18,6 +18,10 @@ pub struct GraphQLArgs {
     #[clap(long = "graphql-number-of-threads", default_value = "2", env)]
     pub graphql_number_of_threads: usize,
 
+    /// The size of the batch fetched from the database by GraphQL service.
+    #[clap(long = "graphql-database-batch-size", default_value = "100", env)]
+    pub database_batch_size: usize,
+
     /// The max depth of GraphQL queries.
     #[clap(long = "graphql-max-depth", default_value = "16", env)]
     pub graphql_max_depth: usize,