+++ crates/fuel-core/src/query/coin.rs
@@ -3,6 +3,7 @@ use fuel_core_storage::{
     iter::IterDirection,
     not_found,
     tables::Coins,
+    Error as StorageError,
     Result as StorageResult,
     StorageAsRef,
 };
@@ -14,6 +15,7 @@ use fuel_core_types::{
 use futures::{
     Stream,
     StreamExt,
+    TryStreamExt,
 };
 
 impl ReadView {
@@ -29,6 +31,18 @@ impl ReadView {
         Ok(coin.uncompress(utxo_id))
     }
 
+    pub async fn coins(
+        &self,
+        utxo_ids: Vec<UtxoId>,
+    ) -> impl Iterator<Item = StorageResult<Coin>> + '_ {
+        // TODO: Use multiget when it's implemented.
+        //  https://github.com/FuelLabs/fuel-core/issues/2344
+        let coins = utxo_ids.into_iter().map(|id| self.coin(id));
+        // Give a chance to other tasks to run.
+        tokio::task::yield_now().await;
+        coins
+    }
+
     pub fn owned_coins(
         &self,
         owner: &Address,
@@ -36,11 +50,17 @@ impl ReadView {
         direction: IterDirection,
     ) -> impl Stream<Item = StorageResult<Coin>> + '_ {
         self.owned_coins_ids(owner, start_coin, direction)
-            .map(|res| {
-                res.and_then(|id| {
-                    // TODO: Move fetching of the coin to a separate thread (https://github.com/FuelLabs/fuel-core/pull/2340)
-                    self.coin(id)
-                })
+            .chunks(self.batch_size)
+            .map(|chunk| {
+                use itertools::Itertools;
+
+                let chunk = chunk.into_iter().try_collect::<_, Vec<_>, _>()?;
+                Ok::<_, StorageError>(chunk)
             })
+            .try_filter_map(move |chunk| async move {
+                let chunk = self.coins(chunk).await;
+                Ok(Some(futures::stream::iter(chunk)))
+            })
+            .try_flatten()
     }
 }